+++ src/_pytask/nodes.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 import functools
 import itertools
+import uuid
 from abc import ABCMeta
 from abc import abstractmethod
 from pathlib import Path
@@ -21,6 +22,7 @@ from _pytask.exceptions import NodeNotCollectedError
 from _pytask.exceptions import NodeNotFoundError
 from _pytask.mark_utils import remove_markers_from_func
 from _pytask.session import Session
+from pybaum import tree_map
 
 
 if TYPE_CHECKING:
@@ -119,8 +121,6 @@ class PythonFunctionTask(MetaTask):
     """Optional[List[Mark]]: A list of markers attached to the task function."""
     kwargs = attr.ib(factory=dict, type=Dict[str, Any])
     """Dict[str, Any]: A dictionary with keyword arguments supplied to the task."""
-    keep_dict = attr.ib(factory=dict, type=Dict[str, bool])
-    """Dict[str, bool]: Should dictionaries for single nodes be preserved?"""
     _report_sections = attr.ib(factory=list, type=List[Tuple[str, str, str]])
     """List[Tuple[str, str, str]]: Reports with entries for when, what, and content."""
     attributes = attr.ib(factory=dict, type=Dict[Any, Any])
@@ -135,17 +135,13 @@ class PythonFunctionTask(MetaTask):
         cls, path: Path, name: str, function: Callable[..., Any], session: Session
     ) -> PythonFunctionTask:
         """Create a task from a path, name, function, and session."""
-        keep_dictionary = {}
-
         objects = _extract_nodes_from_function_markers(function, depends_on)
-        nodes, keep_dict_de = _convert_objects_to_node_dictionary(objects, "depends_on")
-        keep_dictionary["depends_on"] = keep_dict_de
-        dependencies = _collect_nodes(session, path, name, nodes)
+        nodes = _convert_objects_to_node_dictionary(objects, "depends_on")
+        dependencies = tree_map(lambda x: _collect_node(session, path, name, x), nodes)
 
         objects = _extract_nodes_from_function_markers(function, produces)
-        nodes, keep_dict_prod = _convert_objects_to_node_dictionary(objects, "produces")
-        keep_dictionary["produces"] = keep_dict_prod
-        products = _collect_nodes(session, path, name, nodes)
+        nodes = _convert_objects_to_node_dictionary(objects, "produces")
+        products = tree_map(lambda x: _collect_node(session, path, name, x), nodes)
 
         markers = [
             marker
@@ -167,7 +163,6 @@ class PythonFunctionTask(MetaTask):
             produces=products,
             markers=markers,
             kwargs=kwargs,
-            keep_dict=keep_dictionary,
         )
 
     def execute(self, **kwargs: Any) -> None:
@@ -216,8 +211,8 @@ class FilePathNode(MetaNode):
             return str(self.path.stat().st_mtime)
 
 
-def _collect_nodes(
-    session: Session, path: Path, name: str, nodes: dict[str, str | Path]
+def _collect_node(
+    session: Session, path: Path, name: str, node: str | Path
 ) -> dict[str, MetaNode]:
     """Collect nodes for a task.
 
@@ -243,21 +238,16 @@ def _collect_nodes(
         If the node could not collected.
 
     """
-    collected_nodes = {}
-
-    for node_name, node in nodes.items():
-        collected_node = session.hook.pytask_collect_node(
-            session=session, path=path, node=node
+    collected_node = session.hook.pytask_collect_node(
+        session=session, path=path, node=node
+    )
+    if collected_node is None:
+        raise NodeNotCollectedError(
+            f"{node!r} cannot be parsed as a dependency or product for task "
+            f"{name!r} in {path!r}."
         )
-        if collected_node is None:
-            raise NodeNotCollectedError(
-                f"{node!r} cannot be parsed as a dependency or product for task "
-                f"{name!r} in {path!r}."
-            )
-        else:
-            collected_nodes[node_name] = collected_node
 
-    return collected_nodes
+    return collected_node
 
 
 def _extract_nodes_from_function_markers(
@@ -277,80 +267,50 @@ def _extract_nodes_from_function_markers(
         yield parsed
 
 
-def _convert_objects_to_node_dictionary(
-    objects: Any, when: str
-) -> tuple[dict[Any, Any], bool]:
+def _convert_objects_to_node_dictionary(objects: Any, when: str) -> dict[Any, Any]:
     """Convert objects to node dictionary."""
-    list_of_tuples, keep_dict = _convert_objects_to_list_of_tuples(objects, when)
-    _check_that_names_are_not_used_multiple_times(list_of_tuples, when)
-    nodes = _convert_nodes_to_dictionary(list_of_tuples)
-    return nodes, keep_dict
+    list_of_dicts = [convert_to_dict(x) for x in objects]
+    _check_that_names_are_not_used_multiple_times(list_of_dicts, when)
+    nodes = merge_dictionaries(list_of_dicts)
+    return nodes
 
 
-def _convert_objects_to_list_of_tuples(
-    objects: Any | tuple[Any, Any] | list[Any] | list[tuple[Any, Any]], when: str
-) -> tuple[list[tuple[Any, ...]], bool]:
-    """Convert objects to list of tuples.
+@attr.s(frozen=True)
+class _Placeholder:
+    scalar = attr.ib(type=bool, default=False)
+    id_ = attr.ib(factory=uuid.uuid4, type=uuid.UUID)
 
-    Examples
-    --------
-    _convert_objects_to_list_of_tuples([{0: 0}, [4, (3, 2)], ((1, 4),))
-    [(0, 0), (4,), (3, 2), (1, 4)], False
 
-    """
-    keep_dict = False
-
-    out = []
-    for obj in objects:
-        if isinstance(obj, dict):
-            obj = obj.items()
-
-        if isinstance(obj, Iterable) and not isinstance(obj, str):
-            keep_dict = True
-            for x in obj:
-                if isinstance(x, Iterable) and not isinstance(x, str):
-                    tuple_x = tuple(x)
-                    if len(tuple_x) in [1, 2]:
-                        out.append(tuple_x)
-                    else:
-                        name = "Dependencies" if when == "depends_on" else "Products"
-                        raise ValueError(
-                            f"{name} in pytask.mark.{when} can be given as a value or "
-                            "a name and a value which is 1 or 2 elements. The "
-                            f"following node has {len(tuple_x)} elements: {tuple_x}."
-                        )
-                else:
-                    out.append((x,))
+def convert_to_dict(x: Any, first_level: bool = True) -> Any | dict[Any, Any]:
+    if isinstance(x, dict):
+        return {k: convert_to_dict(v, False) for k, v in x.items()}
+    elif isinstance(x, Iterable) and not isinstance(x, str):
+        if first_level:
+            return {
+                _Placeholder(): convert_to_dict(element, False)
+                for i, element in enumerate(x)
+            }
         else:
-            out.append((obj,))
-
-    if len(out) > 1:
-        keep_dict = False
-
-    return out, keep_dict
+            return {i: convert_to_dict(element, False) for i, element in enumerate(x)}
+    elif first_level:
+        return {_Placeholder(scalar=True): x}
+    else:
+        return x
 
 
 def _check_that_names_are_not_used_multiple_times(
-    list_of_tuples: list[tuple[Any, ...]], when: str
+    list_of_dicts: list[dict[Any, Any]], when: str
 ) -> None:
     """Check that names of nodes are not assigned multiple times.
 
     Tuples in the list have either one or two elements. The first element in the two
     element tuples is the name and cannot occur twice.
 
-    Examples
-    --------
-    >>> _check_that_names_are_not_used_multiple_times(
-    ...     [("a",), ("a", 1)], "depends_on"
-    ... )
-    >>> _check_that_names_are_not_used_multiple_times(
-    ...     [("a", 0), ("a", 1)], "produces"
-    ... )
-    Traceback (most recent call last):
-    ValueError: '@pytask.mark.produces' has nodes with the same name: {'a'}
-
     """
-    names = [x[0] for x in list_of_tuples if len(x) == 2]
+    names_with_provisional_keys = list(
+        itertools.chain.from_iterable(dict_.keys() for dict_ in list_of_dicts)
+    )
+    names = [x for x in names_with_provisional_keys if not isinstance(x, _Placeholder)]
     duplicated = find_duplicates(names)
 
     if duplicated:
@@ -359,37 +319,63 @@ def _check_that_names_are_not_used_multiple_times(
         )
 
 
-def _convert_nodes_to_dictionary(
-    list_of_tuples: list[tuple[Any, ...]]
-) -> dict[Any, Any]:
-    """Convert nodes to dictionaries.
+def union_of_dictionaries(dicts: list[dict[Any, Any]]) -> dict[Any, Any]:
+    """Merge multiple dictionaries in one.
 
     Examples
     --------
-    >>> _convert_nodes_to_dictionary([(0,), (1,)])
-    {0: 0, 1: 1}
-    >>> _convert_nodes_to_dictionary([(1, 0), (1,)])
-    {1: 0, 0: 1}
+    >>> a, b = {"a": 0}, {"b": 1}
+    >>> union_of_dictionaries([a, b])
+    {'a': 0, 'b': 1}
+
+    >>> a, b = {'a': 0}, {'a': 1}
+    >>> union_of_dictionaries([a, b])
+    {'a': 1}
 
     """
-    nodes = {}
-    counter = itertools.count()
-    names = [x[0] for x in list_of_tuples if len(x) == 2]
+    return dict(itertools.chain.from_iterable(dict_.items() for dict_ in dicts))
 
-    for tuple_ in list_of_tuples:
-        if len(tuple_) == 2:
-            node_name, node = tuple_
-            nodes[node_name] = node
 
+def merge_dictionaries(list_of_dicts: list[dict[Any, Any]]) -> dict[Any, Any]:
+    """Merge multiple dictionaries.
+
+    The function does not perform a deep merge. It simply merges the dictionary based on
+    the first level keys which are either unique names or placeholders. During the merge
+    placeholders will be replaced by an incrementing integer.
+
+    Examples
+    --------
+    >>> a, b = {"a": 0}, {"b": 1}
+    >>> merge_dictionaries([a, b])
+    {'a': 0, 'b': 1}
+
+    >>> a, b = {_Placeholder(): 0}, {_Placeholder(): 1}
+    >>> merge_dictionaries([a, b])
+    {0: 0, 1: 1}
+
+    """
+    merged_dict = union_of_dictionaries(list_of_dicts)
+
+    if len(merged_dict) == 1 and isinstance(list(merged_dict)[0], _Placeholder):
+        placeholder, value = list(merged_dict.items())[0]
+        if placeholder.scalar:
+            out = value
         else:
-            while True:
-                node_name = next(counter)
-                if node_name not in names:
-                    break
+            out = {0: value}
+    else:
+        counter = itertools.count()
+        out = {}
+        for k, v in merged_dict.items():
+            if isinstance(k, _Placeholder):
+                while True:
+                    possible_key = next(counter)
+                    if possible_key not in merged_dict and possible_key not in out:
+                        out[possible_key] = v
+                        break
+            else:
+                out[k] = v
 
-            nodes[node_name] = tuple_[0]
-
-    return nodes
+    return out
 
 
 def create_task_name(path: Path, base_name: str) -> str: