+++ tests/test_task.py
@@ -8,7 +8,7 @@ from pytask import ExitCode
 from pytask import main
 
 
-@pytest.mark.unit
+@pytest.mark.end_to_end
 @pytest.mark.parametrize("func_name", ["task_example", "func"])
 @pytest.mark.parametrize("task_name", ["the_only_task", None])
 def test_task_with_task_decorator(tmp_path, func_name, task_name):
@@ -37,7 +37,7 @@ def test_task_with_task_decorator(tmp_path, func_name, task_name):
         )
 
 
-@pytest.mark.unit
+@pytest.mark.end_to_end
 @pytest.mark.parametrize("func_name", ["task_example", "func"])
 @pytest.mark.parametrize("task_name", ["the_only_task", None])
 def test_task_with_task_decorator_with_parametrize(tmp_path, func_name, task_name):